refactor: Move code into src directory
- Move crawler.py and website_crawler.py into src/
- Update imports to use src directory
- Update pyproject.toml to include src package
